new changed

+++ second.py
@@ -1,4 +1,7 @@
 import random
+import sys
+import os
+import urllib.parse
 
 def test_for_inclusion(name,namefile):
     with open(namefile,'r') as myfile:
@@ -27,34 +30,39 @@ def new_names(firstname,namefile):
 
 
 def main():
+
+    if len(sys.argv) == 1:
+        namefile='list_names.txt'
+    else:
+        namefile=sys.argv[1]
+        
+    if not os.path.exists(namefile):
+        filefirst=open(namefile,'w')
+        filefirst.close()
+    
     firstname=input('Введите имя\n')
-    namefile='list_names.txt'
-    try:
-        filefirst=open(namefile,'r')
-        exept :
-            filefirst=open(namefile,'w')
 
-            doubling=test_for_inclusion(firstname,namefile)
-            print(doubling)
-            doubling0=0
+    doubling=test_for_inclusion(firstname,namefile)
+    print(doubling)
+    doubling0=0
 
-            if doubling==1:
-            names=new_names(firstname,namefile)
-            print("Такое имя уже существует. Выберите из предложенных\n",names)
-            firstname=input()
-            for element in names:
-                if firstname==element:
-                    doubling0=0
-                    break
-                else:
-                    doubling0=1
-
-            if doubling0==0:
-                with open('list_names.txt','a') as myfile:
-                    print (firstname, file=myfile)
-                    print('Пользователь зарегистрирован')
+    if doubling==1:
+        names=new_names(firstname,namefile)
+        print("Такое имя уже существует. Выберите из предложенных\n",names)
+        firstname=input()
+        for element in names:
+            if firstname==element:
+                doubling0=0
+                break
             else:
-                print('Пользователь НЕ зарегестрирован')
+                doubling0=1
+
+    if doubling0==0:
+        with open('list_names.txt','a') as myfile:
+            print (firstname, file=myfile)
+            print('Пользователь зарегистрирован')
+    else:
+        print('Пользователь НЕ зарегестрирован')
                 
 
 if __name__=='__main__':
